requirements.txt
numpy>=1.20.0
pandas>=1.3.0
matplotlib>=3.4.0
pyyaml>=6.0
yfinance>=0.2.18
mplfinance>=0.12.9b0
requests>=2.28.0
python-dateutil>=2.8.2
scipy>=1.7.0
alpaca-py>=0.8.0
TA-Lib>=0.4.24
flask>=2.2.0
beautifulsoup4>=4.11.0
gunicorn>=20.1.0
werkzeug>=2.2.0
flask-cors>=3.0.10